Remove variable store.
It was introduced to determine what the domain of a variable is, thinking that it is not the case that all domains in the engine match the domains from the model. However, in the search component we're only dealing with search variables, and their domains will match the domains from the model.