Feature/neighbourhood (#68)
* Fix compilation errors in the assignment, and mark probe as const.

* Introduce the move and the base neighbourhood class.

* Introduce the variable store, which provides metadata on variables to the search.

As of now, this metadata is just the bounds of the domain in the engine. However, for some variables the domain in the engine is larger than defined in the FlatZinc model. This would be the place to abstract that.

* Implement random neighbourhood.

* Implement actual domains for the variable store.

* Remove variable store.

It was introduced to determine what the domain of a variable is, thinking that it is not the case that all domains in the engine match the domains from the model. However, in the search component we're only dealing with search variables, and their domains will match the domains from the model.

* Change the design of the Move class.

* Made RandomProvider::element assume the collection is non-empty.

* Redesign the neighbourhood interface to localise the template argument for Move.

* Make Annealer::acceptMove() public, and remove unused parameter.

* Rename AssignmentModification to AssignmentModifier.