Feature/gcc (#71)
* Added the GCC constraint

* GCC Open/Close is now a template bool

The openness of a GCC constraint is now a template bool.
GCC is now up-to-date with the refactoring of the merge.
Additional unit tests need to be added for the GCC

* Added the GCC constraint

* GCC Open/Close is now a template bool

The openness of a GCC constraint is now a template bool.
GCC is now up-to-date with the refactoring of the merge.
Additional unit tests need to be added for the GCC

* updated gcc

fixed bugs. Added more tests.